+++ QCut/qpd/__init__.py
@@ -1,0 +1,1 @@
+"""Quasi-probability decompositions of the cut gates and wires."""